refactor: оптимизировал код BreadCrumbs и Crumb
