+++ src/components/BreadCrumbs/index.tsx
@@ -4,7 +4,6 @@ import { useLocation } from 'react-router-dom';
 
 import './BreadCrumbs.scss';
 
-import { menuItems } from '../../utils/constants';
 import { useSelector } from 'react-redux';
 import { selectNewsData } from '../../redux/news/selectors';
 import { selectUnionMembersData } from '../../redux/unionMember/selectors';
@@ -12,13 +11,10 @@ import { selectProjectsData } from '../../redux/project/selectors';
 import { selectVideoData } from '../../redux/video/selectors';
 import { selectArticleData } from '../../redux/article/selectors';
 import { Crumb } from '../Crumb';
-
-export interface CrumbTexts {
-  [key: string]: string;
-}
+import { crumbTexts } from '../../utils/crumbTexts';
 
 export interface Item {
-  _id?: string;
+  _id: string;
   imageUrl?: string;
   title?: string;
   name?: string;
@@ -38,20 +34,9 @@ export const BreadCrumbs = () => {
   const { item: articleItem } = useSelector(selectArticleData);
 
   const { pathname } = useLocation();
+  
   const pathnames = pathname.split('/').filter((pathname) => pathname);
   pathnames.unshift('/');
-
-  const crumbTexts: CrumbTexts = {
-    '/': menuItems[0].name,
-    news: menuItems[1].name,
-    unionmembers: menuItems[2].name,
-    projects: menuItems[3].name,
-    scores: menuItems[4].name,
-    media: menuItems[5].name,
-    reports: menuItems[6].name,
-    aboutus: menuItems[7].name,
-    contacts: menuItems[8].name,
-  };
 
   const item: Item = pathname.includes('news')
     ? newsItem
@@ -63,15 +48,11 @@ export const BreadCrumbs = () => {
     ? articleItem
     : videoItem;
 
-  return (
-    <nav className='bread-crumbs'>
-      {pathnames.length > 0 && (
-        <ul className='bread-crumbs__list'>
-          {pathnames.map((pathname, index) => (
-            <Crumb item={item} index={index} pathname={pathname} pathnames={pathnames} crumbTexts={crumbTexts} />
-          ))}
-        </ul>
-      )}
-    </nav>
-  );
+  const crumbs = pathnames.map((pathname, index) => (
+    <li key={index}>
+      <Crumb item={item} index={index} pathname={pathname} pathnames={pathnames} crumbTexts={crumbTexts} />
+    </li>
+  ));
+
+  return <nav className='bread-crumbs'>{<ul className='bread-crumbs__list'>{crumbs}</ul>}</nav>;
 };
