+++ src/components/Crumb/index.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { CrumbTexts, Item } from '../BreadCrumbs';
+import { Item } from '../BreadCrumbs';
+import { CrumbTexts } from '../../utils/crumbTexts';
+
 import './Crumb.scss';
 
 type CrumbTextProps = {
@@ -12,22 +14,17 @@ type CrumbTextProps = {
 };
 
 export const Crumb = ({ item, index, pathnames, crumbTexts, pathname }: CrumbTextProps) => {
-  return (
+  const crumbElement = (
     <>
-      {index === pathnames.length - 1 ? (
-        <li className='crumb' key={index}>
-          {pathnames.length === 3
-            ? item.title || `${item.surname} ${item.name} ${item.patronymic}`
-            : crumbTexts[pathname]}
-        </li>
-      ) : (
-        <li className='crumb' key={index}>
-          <Link className='crumb__link' to={`${pathnames.slice(0, index + 1).join('')}`}>
-            {crumbTexts[pathname]}
-          </Link>
-          <span className='crumb__span'>|</span>
-        </li>
-      )}
+      <Link className='crumb__link' to={`${pathnames.slice(0, index + 1).join('')}`}>
+        {crumbTexts[pathname]}
+      </Link>
+      <span className='crumb__span'>|</span>
     </>
   );
+
+  const lastCrumbEelement =
+    pathnames.length === 3 ? item.title || `${item.surname} ${item.name} ${item.patronymic}` : crumbTexts[pathname];
+
+  return <div className='crumb'>{index === pathnames.length - 1 ? lastCrumbEelement : crumbElement}</div>;
 };
